Implement simple load balancing and error handling
The Overpass API requests are now randomly distributed across multiple servers.
If a request fails a new request will be started after 1 second for the current camera position. This process repeats until the current camera position was successfully retrieved. This ensures that at some point in the future the cells/stops in the current users viewport will be loaded, while requests for other "cells" that failed before and are outside of the current camera view won't be re-requested.
Some additional comments and minor corrections were also done in this commit.