Exclude platform nodes that are part of a platform
In some regions it is pretty common to map the pole in addition to the platform. In these situations some mappers add the public_transport = platform tag to the pole. Example: https://www.openstreetmap.org/node/3378927728
In such a case we favor th platform way over the platform node. This change prevents asking the same questions for the pole again.